+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Updated travis.yml to support Ubuntu 18.04, gcc-7, and clang-7.0
 * Contributors guidelines updated
+* Avoid cstdlib random generators in ransac registration, use C++11 random instead.
 
 ## 0.9.0